improve dark mode, login/signup redirect
