+++ src/App.tsx
@@ -4,8 +4,8 @@ import { RecoilRoot } from 'recoil';
 import { Routes, Route } from 'react-router-dom';
 import { ThemeProvider } from '@mui/material';
 import useGaTracker from 'hooks/useGaTracker';
-import LoginPage from 'components/LoginPage/LoginPage';
-import SignUpPage from 'components/SignupPage/SignUpPage';
+import LoginPage from 'pages/LoginPage';
+import SignUpPage from 'pages/SignUpPage';
 import theme from 'mui-theme/theme';
 import Index from 'pages/index';
 import NotFoundPage from 'pages/404';
