+++ src/components/Dialog/Dialog.tsx
@@ -4,17 +4,17 @@ import { Dialog as MUIDialog, DialogTitle, Grid, IconButton, Typography, useMedi
 import CloseIcon from '@mui/icons-material/Close';
 
 import Container from 'components/Container/Container';
-// import { useGlobalContext } from 'GlobalContext';
+import { useGlobalContext } from 'darkModeContext';
 const Dialog = ({ onClose, title, content, }: { onClose: () => void; title: string; content: ReactNode;  }) => {
   const theme = useTheme();
   const fullScreen = useMediaQuery(theme.breakpoints.down('md'));
-  // const { darkMode} = useGlobalContext()
+  const { darkMode} = useGlobalContext()
   return (
     <MUIDialog key='settings_popup' fullScreen={fullScreen} maxWidth='md' fullWidth open >
-      <Container isSecondary>
+      <Container isSecondary={darkMode}>
         <DialogTitle>
           <Grid container direction='row' justifyContent='space-between' alignItems='center'>
-            <Typography variant='h4' component='span' color='#0d174d' sx={{ flexGrow: 1, fontWeight: 'bold' }} >
+            <Typography variant='h4' component='span' color= {darkMode? '#0d174d':'white'} sx={{ flexGrow: 1, fontWeight: 'bold' }} >
               {title}
             </Typography>
             <IconButton aria-label='close' color='secondary' onClick={onClose}>
