+++ src/pages/About.tsx
@@ -4,23 +4,26 @@ import { Grid, Link, Typography } from '@mui/material';
 import CoffeeIcon from '@mui/icons-material/Coffee';
 import PaidIcon from '@mui/icons-material/Paid';
 import GitHubIcon from '@mui/icons-material/GitHub';
-
+import { useGlobalContext } from 'darkModeContext';
 import Dialog from 'components/Dialog/Dialog';
 
-const About = ({ onClose }: { onClose: () => void }) => (
+const About = ({ onClose }: { onClose: () => void }) => {
+  const { darkMode} = useGlobalContext()
+  return(
+  
   <Dialog
   
     onClose={onClose}
     title='About the App & Me'
     content={
       <Grid container spacing={0}>
-        <Grid item xs={12} style={{ padding: 32, paddingTop: 0 }}>
+        <Grid item xs={12} style={{ padding: 32, paddingTop: 0, color: darkMode? 'black':'white' }}>
           <Typography variant='h4' align='center' component='p' color='secondary'>
             HIIT timer
           </Typography>
 
           <br />
-          <Typography variant='body1' component='p'>
+          <Typography variant='body1' component='p' >
             Timer-set to use during your HIIT trainings, or sprints or... whatever you need!
           </Typography>
           <Typography variant='body1' component='p'>
@@ -84,5 +87,6 @@ const About = ({ onClose }: { onClose: () => void }) => (
     }
   />
 );
+  }
 
 export default memo(About);
